refactor: move and rename tests
